Add data collection attachment endpoints
